refactor(memory): split query/smart_walk.rs (1695 lines) into smart_walk/ directory
Splits the monolithic smart_walk.rs into focused sub-modules:
- types.rs: public types, constants, and truncate_chars helper
- prompts.rs: system/tool prompts, content inventory, model resolution, tool-call parser, fallback synthesis + InnerCall struct
- dispatch.rs: dispatch_call router + all per-action dispatch_* functions
- runner.rs: run_smart_walk main agentic loop
- tool.rs: SmartMemoryWalkTool + ChatProviderAdapter
- smart_walk_tests.rs: extracted inline #[cfg(test)] mod
- mod.rs: re-exports preserving the external API unchanged
